cryptohome: Apply tidy fixes for modernize-concat-nested-namespaces
I'm not entirely sure about this style, although that's probably because
I don't see it as that big of an improvement or detriment either way.
However, it is the lightly-preferred way to do nested namespaces in
Google style so it doesn't hurt to adopt it (and clang-tidy did most of
the work).

BUG=b:326084107
TEST=unit tests
TEST=clang-tidy

Change-Id: Ifa5a4ee1f022673cfbf2bc4d4971550f460928e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5315399
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>