debugd: Don't encode error strings with base64
Some log entries, e.g. cros_ec_panicinfo are encoded with base64.
When log_tool failes to to read it, it returns an error string that
is put into report. The problem here is that the string is also encoded
with base64, which is somewhat confusing.
Fix this by changing the return value of log parsing functions,
leveraging base::expected to inform the caller that the parsing has
failed, and then in case of failure don't encdode the error.

BUG=b:326359702
TEST=generate_logs and check the output of cros_ec_panicinfo.

Change-Id: I185bd1722a5282302c7d6ada9a5ed7685e864aba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5317389
Commit-Queue: Kornel Dulęba <korneld@google.com>
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>