shill: Refactor P2PDevice creation to allow async response
Update the device creation flow in P2PManager and WiFiProvider to allow
WiFiProvider to create the P2PDevice asynchronously. This is necessary
for Multi-Interface Concurrency since WiFiProvider may be required to
perform slow operations like deleting existing interfaces before it
creates the P2PDevice object.

The new flow is P2PManager calls WiFiProvider::RequestP2PDeviceCreation,
which triggers WiFiProvider to create the P2PDevice and send it to
P2PManager whenever it is ready via P2PManager::OnDeviceCreated.
WiFiProvider makes no guarantee on the sequencing of the its call to
P2PManager::OnDeviceCreated, so P2PManager must be prepared for this to
happen synhcronously or asynchronously. Because Multi-Interface
Concurrency is not yet implemented, the asynchronous call is always
dispatched immediately when the request is made.

P2PManager also now exposes a DeviceCreationFailed() which allows
WiFiProvider to indicate that it failed to create the device.

Some cleanup of dependencies was required to avoid a circular dependency
between WiFiPhy and WiFiProvider. Several classes where only importing
WiFiProvider via WiFiPhy, so those had to be fixed to directly import
WiFiProvider.

BUG=b:328072809
TEST=FEATURES=test emerge-brya shill
TEST=Deploy and test P2P device creation flows still work as expected.

Change-Id: I9ba2b120fba61e961c022389003e17c72e66785f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368665
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>