shill: modb: add explicit attach APN for Verizon
If the cellular modem uses the blank APN in the network attach
settings, the network may assign as attach APN the "vzwadmin" one
instead of the expected "ims" one. The "ims" APN is required for
devices with any IMS capabilities, which in the case of ChromeOS is
exclusively for SMS over IMS support.

Being incorrectly attached to "vzwadmin" has the unwanted side effect
that it seems to break the OTA class 3 provisioning in certain cases.

By explicitly asking for "ims" as attach APN, we ensure "vzwadmin" is
not used as attach.

BUG=b:319149494
TEST=Manual tests with static/dynamic IP VZW SIM cards

Cq-Depend: chromium:5381841
Change-Id: I8856ff60c92d7dbe9b40784370771f346f0dab23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5340131
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>