cryptohome: Standardize HomeDirs on obfuscated username.
This API mostly uses the obfuscated username, but not entirely. This
standardizes it, to simplify and add consistency.

This change is a first step to trying to somewhat separate the username
and directory/device handling aspects of the class, so that we can also
use this to handle "trash" homedirs and implement a somewhat more atomic
Remove operation.

This also de-virtualizes a bunch of HomeDirs functions which are never
overridden. This isn't really necessary, but I think it makes it a
little more readable to see what actually is (currently) mockable.

BUG=b:289892544
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I49ad49727be1bf2a65dfd7f02cecb0359d94085a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5319649
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>