shill: Use CreateP2PGroup helper in more P2PManager tests
P2PManager unit tests have a helper function that is used to create a
P2P group. This helps keep the code cleaner and easier to update when
the group creation process changes.

Add new functionality to this function so that it can be used by more
test cases. Adds a CreateP2PGroupWithProperties so that callers can seed
the request with their own desired properties instead of the defaults.
Update the function ro return a KeyValueStore containing the CreateGroup
callback response dictionary so that callers can validate the callback
response.

BUG=b:328072809
TEST=FEATURES=test emerge-brya shill

Change-Id: I20ea256cd27d0d5404dff8da081a704580e8ef97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368663
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>