shill: Add helpers for connecting to P2P groups in P2PManager tests
Add helper functions ConnectToP2PGroupWithProperties and
ConnectToP2PGroup to P2PManager unit tests. The group connection process
is expected to change a lot over the course of P2P development, so it's
helpful to keep the logic in one place.

The test case MissingArgs_ConnectClient does not use these helpers
because it uses a separate flow and expects the command to fail.

BUG=b:328072809
TEST=FEAUTES=test emerge-brya shill

Change-Id: Ib6bdb4846d3e50f788c0bf701601830f42fcea82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5368664
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>