libbrillo: Add obfuscated username overloads for cryptohome path lookups
This adds overloads for several functions used to compute cryptohome
paths, allowing them to be used with both plain and obfuscated
usernames. This allows the functions to be used in code without
requiring that code to have access to the original name.

The overload for GetUserPath duplicates what GetHashedUserPath does.
However, this sort of thing is actually a good candidate for using
overloads: the semantics of both functions are the same, they just
operate on different input types. The original GetHashedUserPath was
written back when these types were just std::string but we no longer
need to do that. I'll send some separate changes to convert the callers
and then remove GetHashedUserPath later.

BUG=b:289892544
TEST=FEATURES=test emerge-$BOARD libbrillo

Change-Id: I1927ea1c1f436528dfbc95e814aaa01cc88916cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5319651
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>