chromeos-config: Run tests in fullcheckout-presubmit
Until we have a well-lit path developed for packages to specify a
simple unit test command, just run as a part of
fullcheckout-presubmit.

It's slightly frowned upon, but we limit ourselves so we don't run on
pre-upload, and we do intend to migrate this to the well-lit path once
available.

The repohooks PRESUBMIT.cfg is another place we do this (and was the
inspiration for this implementation).

BUG=b:326285420
TEST=fullcheckout-presubmit passes

Change-Id: I3df8157ac1d118abac4a8185df9c80cdeb006e12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5319657
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>