lorgnette: Prevent SANE discovery while device is open
When a scanner is open by one client and a second client tries to
perform a discovery, the discovery may corrupt memory for the scanner
that is open.  To prevent this, if a discovery is requested while a
scanner is open, return SANE devices from the cache instead of
performing a new SANE discovery.  This does not affect IPPUSB devices,
so those are still discovered as usual.

BUG=b:323256234
TEST=FEATURES=test emerge lorgnette
TEST=Manual test with scanner in lab

Change-Id: If647ac970c0134fbf73b348cec91194fc99a901e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5321287
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Nathan Muggli <nmuggli@google.com>