usb_bouncer: Make AuthorizeAll more error tolerant
Previously AuthorizeAllImpl had some errno filtering that was not
applied uniformly. After this change, the errno filtering is applied
wherever an errno is used to check whether the operation failed or not
with a goal of ignoring cases where devices are removed during the
operation.

BUG=b:172211373
TEST=FEATURES=test emerge-betty usb_bouncer

Change-Id: Icd959808cf8d35de0268738b28a3bf5518d732f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5321297
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>