shill: wifi: handle ANQP updates in BSS.
ANQP information can be fetched using the ANQPGet D-Bus call. The
result is stored in the BSS properties and listeners are notified with
the BSS "PropertiesChanged" signal.

The WiFiEndpoint now handles the "ANQP" field provided with the
"PropertiesChanged" signal. For now, only the "CapabilityList" field is
parsed for future metrics. On parse success the WiFiDevice is notified
with NotifyANQPInformationChanged().

BUG=b:322972968
TEST=FEATURES="test" emerge-$BOARD shill
  Connect to an access point with ANQP fields and check in the logs the
  fields are parsed without error.

Change-Id: I42271d4f549ebd6df1d797291e15956c61b1ea69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5326917
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>