fbpreprocessor: move D-Bus handling out of InputManager
Move the logic handling the D-Bus signal coming from crash-reporter when
a new firmware dumps is available to its own class.
That allows a cleaner separation between the code that is in charge of
parsing D-Bus messages and the code that handles the pseudonymization of
the firmware dumps. Among other things, that allows us to unit test the
latter without having to introduce the complexity that comes with D-Bus.

BUG=b:309365285
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord
TEST=cros deploy ${DUT} fbpreprocessord
TEST=restart fbpreprocessord
TEST=echo 1 > /sys/kernel/debug/iwlwifi/0000\:00\:14.3/iwlmvm/fw_dbg_collect
TEST=ls -lR /run/daemon-store/fbpreprocessord/ shows a firmware dump in
"processed_dumps"
TEST=dbus-send --system --print-reply=literal
--dest=org.chromium.FbPreprocessor /org/chromium/FbPreprocessor
org.chromium.FbPreprocessor.GetDebugDumps returns the expected output
https://paste.googleplex.com/5033126960562176

Change-Id: Ifad152e88d628ebed62d92caefb06bd51ce71b1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5341792
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Yichen Yu <yichenyu@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>