patchpanel: Use precompiled seccomp filter for iptables
We observed that compiling the seccomp filter for iptables may be
CPU-consuming, and since iptables is invoked a lot of times by
patchpanel, letting patchpanel load a precompiled version of the filter
will be helpful.

This patch changes to use install_seccomp_policy template in the gn file
to compile and install the precompiled filter, and load it in the
MinijailProcessRunner. The file structure for the policy files is also
changed since the previous one is not compatible with this template.

On cherry with cros-debug build, this patch can reduce the patchpanel
start time by ~0.3s (from ~4.2s to ~3.9s).

BUG=b:326852409
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=Verified that seccomp filter is working by commenting out some
  rules and checking that some iptables commands are failing.

Cq-Depend: chromium:5320533
Change-Id: Iddc5549ebd5dc211b8a99cb67f1711603ad488a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5320181
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>