missive: Do not ping Chrome until key is received
Instead of starting periodic updates right away, add key request to
Flush (alike was already done in Write), and start periodic updates
after that.

Modify tests to match.

Later on we will add dummy Flush to Chrome upon enrollment too; at the
moment it is not there, and so it can happen that if connection to
server is lost right after enrollment, events will not be able to
enqueue until the connection is re-established. This is rare case that
will be taken care of before the next release.

BUG=b:322199818
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual on VM

Change-Id: I8b95a7c72704198132b431fca6afcd48fe883949
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5325468
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>