lorgnette: Make CancelScan wait for cancel to complete
In SANE, sane_cancel is just a request to cancel.  The actual
cancellation happens some time later (assuming the backend doesn't just
ignore the flag).  For documentScan and the Scan app, the semantics we
want to present are that the cancel has either succeeded or failed after
CancelScan returns.  This can be achieved by internally looping and
reading from the device until SANE_CANCELLED is returned.

This needs to be balanced against the fact that we don't want to block
the d-bus response for any longer than necessary.  Waiting for a scanner
to finish cancelling the job may take as long as waiting for the end of
the current page.  As a compromise, make CancelScan wait up to 3s for
the cancel to complete, then return DEVICE_BUSY if it is still in
progress.  The caller can check for this status and retry the request if
needed.

Because waiting for the cancel to complete potentially diverts some of
the document data that the client may be trying to read, we don't want
the client to continue receiving data in its read loop.  Update
ReadScanData to pretend that the read succeeded with no data when a
cancel is in progress.  If a client is sitting in a read loop waiting
for the end of the page, this will eventually return the cancelled
result.

In order to make all this work, job handles need to remain valid long
enough for the client to receive the final status.  Instead of dropping
a job handle as soon as it gets a non-GOOD status, keep them active
until the next call to OpenScanner or StartPreparedScan.  Track the most
recent response with the handle so that it can be returned as many times
as needed.  This avoids having the internal read loop silently swallow
SANE statuses.

Finally, sane-airscan seems to have a bug that propagates a cancelled
status for the ADF from one scan to the next.  It's not completely clear
if this is caused by the way lorgnette is interacting with SANE, but in
the meantime, add a workaround that waits for the end of the current
page before cancelling.  Since the ADF generally can't cancel before it
finishes reading the page,  this shouldn't change the user experience
noticeably.  This does not appear to be needed for other backends, but
can easily be applied to more cases if hardware testing reveals a need.

BUG=b:326587966
TEST=tast documentscanapi.AdvancedScan; test extension on real scanners

Change-Id: I1033cee544f2a4da5ae4ffe543766b61bfbbaba1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5345850
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>