modemfwd: remove unused dbus::Bus param from TryFlash
This is lexically unused, and there's no other reason to have
it around (i.e. ensuring there is a strong reference that is
not held at the point of use for some reason, though that would
also require refactoring), so we can get rid of it.

BUG=None
TEST=unit tests

Change-Id: I0ef06b8379fba47858fd04bcc03e3ce6b07b5b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5333376
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>