modemfwd: Don't lose modem health state on second start call
We call StartHeartbeatTimer when the modem first comes up in
OnModemCarrierIdReady if there is a carrier, or in CheckForWedgedModems
otherwise. If both events happen, we call into StartHeartbeatTimer
twice. This is no issue, except that the second call will clear the
tracked number of heartbeat failures, potentially delaying recovery.

This can happen if the usb device is valid (so wedge detection isn't
needed) but the modem isn't replying over the primary port. Or, if a
test expects a modem recovery within X minutes but wedge detection
resets the count in the middle, then the test will fail (b/327431329)

BUG=b:327431329
TEST=Ran ModemfwdRecovery on a failing FM101 device

Change-Id: I225ac1b36537575abee6d854a5b306d8e105e6b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5333725
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>