power: add ExternalDisplayOnly machine quirk
Leverage the existing machine quirks class and add another function
that allows the ExternalDisplayOnly pref to be set during runtime for
specific machines that match the dmi ids defined in the overlay.

This change only affects a board that has the has-machine-quirk pref
set in its overlay.

BUG=b:316750438
TEST=In chroot:
1) cros_workon_make --board=reven chromeos-base/power_manager --test
2) cros deploy --board=reven $DUT_IP  chromeos-config-bsp chromeos-config
power_manager
On DUT:
3) cros_config /power external-display-only-models
4) confirm that "ExternalDisplayOnly Pref set to enabled" is seen in
/var/log/power_manager/POWERD.LATEST when external-display-only-models
in the overlay is set to the DUT's dmi product_name (and vice versa).

Change-Id: I821398d908dc58e29dfb3b36f1cfe0abd5838028
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5333377
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Esther Shimanovich <eshimanovich@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Esther Shimanovich <eshimanovich@google.com>