chromeos-config: Output a terminating newline when on a tty
cros_config normally dumps the config value with no extra formatting,
but this can lead to weird quirks depending on the terminal emulator
in use (e.g., tmux will cause PS1 to cover the result).

For user convenience, add a terminating newline character when
attached to a tty.  We don't change the behavior when stdout is
attached to something not a terminal, so scripts shouldn't be
affected.

BUG=b:327269772
TEST=CQ

Change-Id: If0e881c2612318d116650237840f5d2efa858f24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5334202
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>