cryptohome: Get rid of the AuthSession bitmask parameter
We want to get rid of these bitmask APIs and just replace them with
explicit flag parameters. I did a first step for this long ago, getting
rid of the flags bitmask internally in AuthSession and pushing the flag
up to AuthSession::Create. This goes to the next level and pushes it all
the way up to the UserDataAuth layer.

I plan to get rid of the bitmask from the dbus API too but that will be
done in other changes because it needs to be done as a sequence of
changes including chrome changes to stop using the existing flags field.
This change just gets rid of all of the usage within cryptohome itself.

BUG=b:266704614
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If51f962ebc079f0e66d966182ddc30199aa75828
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5335774
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>