cryptohome: Implement is_ephemeral_user flag
This implements the is_ephemeral_user flag, to let us eventually remove
the "flags" field from StartAuthSession. To make it work with calls that
use one or the other, we treat the two fields as an OR.

I also converted all of the unit tests over to using the new field,
except for tests that are specifically verifying the existing field
still works.

BUG=b:26704714
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If70f6724c2c3c275457086ea6e98fc8af46886bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5334214
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>