hiberman: Record metrics for image preloading
Record the percentage of encrypted hibernate image data that
could be preloaded into memory until the user logged in, and
also the data rate at which the data was read from storage.

Before this change we read the hibernate image with a single
mmap() call. This worked because we didn't collect information
about the progress and when switching to the hibernated system
the thread would vanish if it hadn't finished preloading yet.
To determine how much data was read until login we need to
split up the single mmap() call into multiple calls reading
chunks and keep of the total. Also we need to stop the thread
if it hasn't completed yet after login, to obtain the stats
we are interested in.

The new struct ImageDataPreloader is responsible for
preloading the image data in a dedicated thread. A channel
is used to ask the thread to abort preloading if needed.

BUG=b:325328140
TEST=chrome://histograms/PreloadStats has reasonable values
  chrome://histograms/ResumeTime looks similar as without
    this change

Change-Id: I7462a5c52adbf91f8c135e8767617bb974c850bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5355499
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Alyssa Haroldsen <kupiakos@google.com>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>