patchpanel: Refactor MinijailedProcessRunner::RunSyncDestroyWithTimeout
Refactor the reading stdout and stderr part of the child process of this
function. Instead of doing the polling by ourselves, redirect the fds to
anonymous files, and read the files after the process exited.

After this change, this function lose the functionality to set timeout
on the execution of the child process. Currently this functionality is
not used by the callers and we can implement it by setting timeout when
waiting child process to exit. Before that, it's better to refactor this
function to a Process class to avoid making a single function too
complicated.

BUG=None
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.TrafficCounters*

Change-Id: I56a87ac90b25adf2cea05b9fffd729ed91e6f1c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5338970
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>