shill: Delay processing scan results until 6GHz scan completes
Previously, after suspend-resume, if foreground scans are turned
off, a DUT would not be able to discover and connect to a 6GHz AP.
After resuming the sequence of events was:
- DUT triggers a legacy band scan
- DUT flushes BSSes that were not found in legacy band scan
- DUT triggers a 6GHz scan
- If a previously connected 6GHz AP is not in the same
  service as a 5GHz AP detected in the legacy band scan, shill
  disconnects due to no remaining endpoints in the current service
- Shill aborts the ongoing 6GHz scan so the DUT cannot find
  the 6GHz AP

In this CL, shill waits until the 6GHz scan completes before
processing scan results to prevent disconnecting too early and
aborting the 6GHz scan. In addition, the scan results are
post-processed so that if a BSS is removed then added back,
this is treated as a property/path change.

BUG=b:335462190
TEST=tast run wifi.SuspendStress.6ghz* should pass. Also ran
a manual test to check that if the 6GHz AP disappears
during suspend, the DUT disconnects as soon as the 6GHz
scan completes.

Change-Id: I5df3d6133f2388d7f6b318b6d6dc21fa9ba6e1df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5338918
Tested-by: Ruth Mekonnen <rmekonnen@google.com>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>