libhwsec: u2fd: Add cert level to GetFipsInfo API
Rename GetFipsStatus API to GetFipsInfo, and include the certification
levels of the implementation along with the response. We need these cert
levels in u2fd, and the best location to deal with data that is
different across GSC types is inside libhwsec.

BUG=b:327104373
TEST=libhwsec/u2fd unittests

Disallow-Recycled-Builds: test-failures
Change-Id: I58565a3eddf0deb32f5b64084057d80dc42b3fef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5339290
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>