shill: portal: avoid HTTP probe that
This patch disables http://www.google.com/gen_204 as a portal detection
probe to avoid asking Chrome to open the portal with a HTTP request to
www.google.com.

Due to a buit-in HSTS policy (different from HSTS preload or
server-controlled HSTS policy), Chrome may automatically try to send an
HTTPS request and inadvertently prevent the portal from redirecting the
request in the sign-in window or tab.

BUG=b:326925490
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: I4f92f60924166d9731bac891476def2af1943859
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5341169
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>