crash: Reduce the number of vmlog lines collected when chrome crashes
VMLOG.LATEST contains some statistics related to cpu and memory usage.
The data is collected every 2s by metrics_daemon.
Currently when chrome crashes 200 latest entries are collected, which
covers the last 400 seconds. I've found that having this much is
somewhat inconvenient when going through the logs.
For chrome crashes the longest we'll ever want is around 60s, which is
the default timeout for session_manager to restart chrome when it's
hung.
Reduce that value to 60, which will cover the last 120s.

BUG=b:324016659
TEST=Crash chrome and check the contents of the *.chrome.txt.gz file

Change-Id: If12b371aca2aad563becda85c99ca0e7d7f3621c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5340018
Commit-Queue: Kornel Dulęba <korneld@google.com>
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>