cryptohome: Block CryptohomeRecoveryAuthBlock::Derive() if locked out
For the PIN and Fingerprint AuthFactor types authentication operation
fails during the key derivation operation in the AuthBlock if the
credential is locked out. This has to be that way since the delay and
the credential state is updated with each wrong attempt.

This makes it difficult to handle the credential lock out in an early
phase during the AuthenticateAuthFactor. Therefore here in this CL we
duplicate the lockout behaviour of PIN and fingerprint AuthBlocks for
CryptohomeRecovery AuthBlock to block the authentication operation
from AuthBlock::Derive() when the crendential is locked.

BUG=b:324076479
TEST=FEATURES=test emerge-betty cryptohome
TEST=tast run DUTVM cryptohome.* hwsec.*

Change-Id: I7ee2356ed3a9fd4ff9b2c53589d43b3d45b3e81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5347155
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Howard Yang <hcyang@google.com>
Tested-by: Betul Soysal <betuls@google.com>