patchpanel: Add iptables batch mode
Add a AcquireIptablesBatchMode() function on the
MinijailedProcessRunner, after it's called, the following iptables() and
ip6tables() calls will be cached and executed together with
`iptables-restore -n` after the returned object is destructed.

This patch only implements this logic and doesn't change any behavior
now. The usage of this function will be added in the follow-up patches.

BUG=b:326495277
TEST=FEATURES="test" emerge-$BOARD patchpanel

Change-Id: Ie99c3bffc624370c33a5795b2657add8932d518f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5330249
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>