debugd: Reap minijail process for ProbeTool
ProbeTool might call itself again in the minijail sub-process in the
past, so we modified it to be non-blocking to avoid deadlock. However,
the change causes the child minijail processes forked by debugd not
being reaped, resulting in the creation of zombie processes. This CL
modifies ProbeTool to post an asynchronous task to reap the child
minijail process created by ProbeTool after releasing it.

BUG=b:328101174
TEST=run `runtime_probe` on DUT, and check for zombie process
TEST=FEATURES=test emerge-dedede debugd

Change-Id: I5c1c065e6c984222a4df807e8dbc929b66468d58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5348399
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Allen Shih <allenshihmc@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Allen Shih <allenshihmc@chromium.org>