attestation: obsolete decrypt
The attestation_client is unable to decrypt what it encrypts. Given that
the Decrypt is not used anywhere AFAIK, we can first obsolete it. In
case we need the Encryption/Decryption in attestation_client in the
future, we can re-implement it in libhwsec. Otherwise, we could remove
the Encryption/Decryption in attestation service/client.

BUG=b:278988633, b:328355827
TEST=unittest

Change-Id: I94511acb28be2ec472b93dfe8a4d1edfd97f79fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5351905
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>