cryptohome: Simplify GetSerializedStringFromToken
This simplifies the function to remove the error path. If you try to
serialize a null token, just return an empty string. This is what the
AuthSession construction code does anyway, with the only other non-test
call site chosing to crash because it never actually passes in a null
token anwyay.

This isn't super critical, I've just been looking through Remove trying
to see if there are ways we can make it more robust and saw an
uninteresting error path we could remove

BUG=b:289892544
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: If740a42b14991ede738e84650193c9e83476a57f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5350109
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>