shill: cellular: Identify Private and Test networks
Instead of hardcoding known private networks in the MODB, identify them
in MobileOperatorInfo based on their MCCMNC.
When a network is a private network, do not schedule a delayed
attach, but start the attach procedure immediately, as private networks
will never be known by the modb.

BUG=b:325318968
TEST=Use callbox_unknown_carrier.pbf to override the modb.
restart shill and verify that the attach is triggered immediately after
OnProfilesChanged is called.

Change-Id: I5408327ad5b920a5a909bfb95b445188ee5e0099
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5349794
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>