login: Print a warning if browser took to long to respond
Log the response time if the browser took at least 1s to respond, and at
least one DBus message has failed. The latter condition is there to make
sure that we've saved the browser state in the logs.
The primary case for this is to have the anomaly detector pick up the
warning message and generate a non-fatal crash.

BUG=b:328017510
TEST=pkill -SIGSTOP chrome; wait for some time; pkill -SIGCONT chrome

Change-Id: I5bb67b1e2f24715e807d4ed2023bfe29d730047f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5352589
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Kornel Dulęba <korneld@google.com>