crash: Watch for browser hangs in anomaly detector
The LivenessChecker mechanism implemented in session_manager is used to
monitor browser responsiveness. The browser is reset after being
unresponsive for 60s, but shorter hangs are also detected and logged.
Extend anomaly_detector to detect those and collect some logs when they
happen. The weight of each event is inversely proportional to the length
of the hang.

BUG=b:328017510
TEST=Hang the browser for some time and verify that a crash is
generated.

Change-Id: Id411cb3f0f8691e1bf97bbbfb3583f7d34b6e4df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5352588
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Marek Maślanka <mmaslanka@google.com>
Commit-Queue: Kornel Dulęba <korneld@google.com>