libstorage: platform: Mount: Use fs type when the source is empty
When calling Mount() for in kernel filesystem, |from| is meaningless; we
use the filesystem type as a convention.
But platform Mount() requires |from| to be absolute, preventing setting
the filesystem type.
Allow empty |from|, use |type| as |from| when the caller send an empty path.

BUG=b:172221451
TEST=Unit tests. Check `mount` output on DUTs:

Change-Id: I3b270d70a26a9f5684f042eb1094242e23dced8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5387035
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>