patchpanel: Delay Manager::Initialize()
The current implementation of Manager::Initialize() is not working as
what we want: the task is not delayed after D-Bus service is ready,
since the D-Bus registration is a sequence of tasks, the PostTask() here
will only insert a task into that sequence but still happen before
patchpanel claims the D-Bus service.

This patch changes the logic to use a PostDelayedTask() instead. See the
inline comment for details. Also change the name to make it clearer and
add some logs to make it easier to debug if there is any issue.

Alternative solution is listening to patchpanel service ready event
inside patchpanel (we don't have other better way to get this signal
now), this will help on the patchpanel startup time but won't help on
the network ready time since it's very likely that the
ConfigureNetwork() request will be handled after Initialize().

This change can help reduce the patchpanel startup time from ~1.2s to
~0.5s on cherry on top of all the other optimizations we have now.

BUG=b:322100632
TEST=CQ

Change-Id: I042b15ae18ee55b2e013aa2baf3c6551cb644e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5355047
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>