camera: Check session parameters to configure streams in FramingSM
In this CL, we enable FramingStreamManipulator by default, and introduce
a condition to skip configuring streams in FramingSM.

ConfigureStreams in FramingSM is skipped if auto framing is not
supported and manual zoom is not requested. Camera clients can request
manual zoom by setting a new vendor tag crosDigitalZoomRequested to
session parameters in the stream config.

BUG=b:314267048
TEST=Check log when opening camera in camera app, Meet

Disallow-Recycled-Builds: test-failures
Change-Id: I35d2215b171b0f514c329ea64f74c818fb47ae6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5354796
Commit-Queue: Kam Kwankajornkiet <kamchonlathorn@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Kam Kwankajornkiet <kamchonlathorn@chromium.org>